Fix prisma settings circular ref
